+++ Src/FM79979Engine/Core/GameplayUT/Log/FMLog.cpp
@@ -106,7 +106,7 @@ namespace FMLog
 #endif
 		if (g_pLatestLog)
 		{
-			if (g_pLatestLog->size() + 1 >= g_iKeepLogCount)
+			if ((int)g_pLatestLog->size() + 1 >= g_iKeepLogCount)
 			{
 				g_pLatestLog->pop_back();
 			}
